Former-commit-id: 7a9da2bf03b4bce6cc3af3187cb7c9b72b0d442a

+++ ifs-data-service/src/main/java/com/worth/ifs/competition/transactional/CompetitionSetupQuestionServiceImpl.java
@@ -164,7 +164,7 @@ public class CompetitionSetupQuestionServiceImpl extends BaseTransactionalServic
                 }
                 formInputRepository.save(appendixFormInput);
             }
-        } else {
+        } else if (appendixFormInput != null) {
             formInputRepository.delete(appendixFormInput);
         }
     }
